// src/state/types.ts
export interface Game {
    title: string
    subtitle: string
    logo: string
    cta: string
    symbol: string
    votes: number
    chain: string
    category?: GameCategories
    address?: string
    twitter?: string
    telegram?: string
    discord?: string
    platform?: GamePlatforms[]
}
// ...
export interface GamesState {
    data: Game[]
    userDataLoaded: boolean
    isLoading: boolean
    prices: {
        address: string
        price: {
            bnb: number
            usd: number
        }
    }[]
}